Change action bar icon back to menu, setting items included. All keyboard buttons aligned and equally stretched.
Former Settings page can also be found in the 'About' page, currently unavailable.
