+++ app/src/main/res/layout/activity_main.xml
@@ -19,6 +19,151 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <include layout="@layout/content_main" />
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="68dp"
+        android:padding="16dp">
+
+        <com.google.android.material.textfield.TextInputLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            android:id="@+id/eq_form_layout"
+            android:minHeight="48dp"
+            android:hint="@string/expression">
+            <com.google.android.material.textfield.TextInputEditText
+                android:id="@+id/eq_form"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:minLines="2"/>
+        </com.google.android.material.textfield.TextInputLayout>
+        <TextView
+            android:id="@+id/mem_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="16sp"
+            app:layout_constraintTop_toBottomOf="@id/eq_form_layout"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="Mem: 0"
+            android:layout_marginTop="8dp"
+            android:layout_marginEnd="4dp"/>
+
+        <TextView
+            android:id="@+id/res_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="20dp"
+            android:textSize="32sp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            tools:text="Result: 0"
+            app:layout_constraintBottom_toTopOf="@id/calc_button" />
+
+        <Button
+            android:id="@+id/calc_button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/calculate"
+            android:layout_marginBottom="20dp"
+            app:layout_constraintBottom_toTopOf="@id/keyboard_grid"/>
+
+        <GridLayout
+            android:id="@+id/keyboard_grid"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:columnCount="4"
+            android:rowCount="8"
+            app:layout_constraintBottom_toTopOf="@id/keyboard_button"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent">
+
+            <View
+                android:id="@+id/separator1"
+                android:layout_width="match_parent"
+                android:layout_height="1dp"
+                android:background="#aac0c0c0"
+                android:layout_marginBottom="20dp"
+                android:layout_marginStart="12dp"
+                android:layout_marginEnd="12dp"
+                android:layout_rowWeight="1"
+                android:layout_columnSpan="4"/>
+
+            <Button
+                android:id="@+id/kb_mc"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="MC"
+                android:textSize="20sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_mp"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="M+"
+                android:textSize="20sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_mm"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="M-"
+                android:textSize="20sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_mr"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="MR"
+                android:textSize="18sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_func"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="func"
+                android:textSize="18sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_const"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#1100aa00"
+                android:text="cons"
+                android:textSize="18sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_c"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#11aa0000"
+                android:text="C"
+                android:textSize="18sp"
+                android:layout_columnWeight="1"/>
+
+            <Button
+                android:id="@+id/kb_back"
+                android:layout_marginStart="5dp"
+                android:backgroundTint="#11aa0000"
+                android:text="←"
+                android:textSize="18sp"
+                android:layout_columnWeight="1"/>
+        </GridLayout>
+
+        <Button
+            android:id="@+id/keyboard_button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="12dp"
+            android:backgroundTint="#00000000"
+            android:text="@string/hide_dedicated_keyboard"
+            android:textColor="?attr/textFillColor"
+            app:layout_constraintBottom_toBottomOf="parent"
+            />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
