+++ app/src/main/res/layout/activity_settings.xml
@@ -1,9 +1,91 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:fitsSystemWindows="true"
     tools:context=".SettingsActivity">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+    <com.google.android.material.appbar.AppBarLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:fitsSystemWindows="true">
+
+        <com.google.android.material.appbar.MaterialToolbar
+            android:id="@+id/toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize" />
+
+    </com.google.android.material.appbar.AppBarLayout>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".SettingsActivity"
+        android:padding="16dp">
+
+        <ImageView
+            android:id="@+id/mode_select_icon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:importantForAccessibility="no"
+            app:layout_constraintBottom_toBottomOf="@id/mode_select"
+            app:layout_constraintEnd_toStartOf="@id/mode_select"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@id/mode_select"
+            app:srcCompat="@drawable/ic_mode_select_outline" />
+
+        <com.google.android.material.switchmaterial.SwitchMaterial
+            android:id="@+id/mode_select"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:minHeight="48dp"
+            android:layout_marginStart="12dp"
+            android:layout_marginTop="64dp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toEndOf="@id/mode_select_icon"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="@string/mode_rad"/>
+
+        <ImageView
+            android:id="@+id/dec_accu_icon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:importantForAccessibility="no"
+            app:layout_constraintBottom_toBottomOf="@+id/accuracy_slider"
+            app:layout_constraintEnd_toStartOf="@+id/slider_desc"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@+id/accuracy_slider"
+            app:srcCompat="@drawable/ic_dec_accu" />
+
+        <TextView
+            android:id="@+id/slider_desc"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/accuracy"
+            android:layout_marginStart="12dp"
+            app:layout_constraintBottom_toBottomOf="@id/accuracy_slider"
+            app:layout_constraintStart_toEndOf="@id/dec_accu_icon"
+            app:layout_constraintTop_toTopOf="@id/accuracy_slider" />
+
+        <com.google.android.material.slider.Slider
+            android:id="@+id/accuracy_slider"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="8dp"
+            android:hapticFeedbackEnabled="true"
+            android:stepSize="1"
+            android:valueFrom="0"
+            android:valueTo="15"
+            android:value="5"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@id/slider_desc"
+            app:layout_constraintTop_toBottomOf="@id/mode_select"
+            app:tickVisible="false" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
