+++ app/src/main/res/layout/fragment_first.xml
@@ -18,7 +18,8 @@
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/eq_form"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:minLines="2"/>
     </com.google.android.material.textfield.TextInputLayout>
     <TextView
         android:id="@+id/mem_text"
@@ -31,106 +32,45 @@
         android:layout_marginTop="8dp"
         android:layout_marginEnd="4dp"/>
 
-    <View
-        android:id="@+id/separator1"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="#aac0c0c0"
-        android:layout_marginTop="32dp"
-        android:layout_marginStart="12dp"
-        android:layout_marginEnd="12dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/mem_text"/>
-
-    <ImageView
-        android:id="@+id/mode_select_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:importantForAccessibility="no"
-        app:layout_constraintBottom_toBottomOf="@id/mode_select"
-        app:layout_constraintEnd_toStartOf="@id/mode_select"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/mode_select"
-        app:srcCompat="@drawable/ic_mode_select_outline" />
-
-    <com.google.android.material.switchmaterial.SwitchMaterial
-        android:id="@+id/mode_select"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:minHeight="48dp"
-        android:layout_marginStart="12dp"
-        android:layout_marginTop="32dp"
-        app:layout_constraintTop_toBottomOf="@id/separator1"
-        app:layout_constraintStart_toEndOf="@id/mode_select_icon"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:text="@string/mode_rad"/>
-
-    <ImageView
-        android:id="@+id/dec_accu_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:importantForAccessibility="no"
-        app:layout_constraintBottom_toBottomOf="@+id/accuracy_slider"
-        app:layout_constraintEnd_toStartOf="@+id/slider_desc"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/accuracy_slider"
-        app:srcCompat="@drawable/ic_dec_accu" />
-
     <TextView
-        android:id="@+id/slider_desc"
+        android:id="@+id/res_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/accuracy"
-        android:layout_marginStart="12dp"
-        app:layout_constraintBottom_toBottomOf="@id/accuracy_slider"
-        app:layout_constraintStart_toEndOf="@id/dec_accu_icon"
-        app:layout_constraintTop_toTopOf="@id/accuracy_slider" />
-
-    <com.google.android.material.slider.Slider
-        android:id="@+id/accuracy_slider"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:hapticFeedbackEnabled="true"
-        android:stepSize="1"
-        android:valueFrom="0"
-        android:valueTo="15"
-        android:value="5"
+        android:layout_marginBottom="20dp"
+        android:textSize="32sp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/slider_desc"
-        app:layout_constraintTop_toBottomOf="@id/mode_select"
-        app:tickVisible="false" />
+        app:layout_constraintStart_toStartOf="parent"
+        tools:text="Result: 0"
+        app:layout_constraintBottom_toTopOf="@id/calc_button" />
 
     <Button
         android:id="@+id/calc_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="24dp"
         android:text="@string/calculate"
-        app:layout_constraintTop_toBottomOf="@id/accuracy_slider"/>
-
-    <TextView
-        android:id="@+id/res_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:textSize="32sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:text="Result: 0"
-        app:layout_constraintTop_toBottomOf="@+id/calc_button" />
+        android:layout_marginBottom="20dp"
+        app:layout_constraintBottom_toTopOf="@id/keyboard_grid"/>
 
     <GridLayout
         android:id="@+id/keyboard_grid"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:columnCount="4"
-        android:rowCount="7"
+        android:rowCount="8"
         app:layout_constraintBottom_toTopOf="@id/keyboard_button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent">
+
+        <View
+            android:id="@+id/separator1"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="#aac0c0c0"
+            android:layout_marginBottom="20dp"
+            android:layout_marginStart="12dp"
+            android:layout_marginEnd="12dp"
+            android:layout_rowWeight="1"
+            android:layout_columnSpan="4"/>
 
         <Button
             android:id="@+id/kb_mc"
