+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,19 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/action_settings"
-        android:icon="@drawable/ic_settings"
-        app:showAsAction="always"
-        android:title="@string/action_settings"/>
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/action_rad_mode"
+            android:checked="true"
+            android:title="RAD" />
+
+        <item
+            android:id="@+id/action_deg_mode"
+            android:title="DEG" />
+    </group>
+
+    <item android:id="@+id/action_accuracy_dialog"
+        android:title="Accuracy"/>
+
+    <item android:id="@+id/action_about"
+        android:title="About"/>
 </menu>
