+++ app/src/main/res/menu/menu_settings.xml
@@ -1,0 +1,7 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item android:id="@+id/action_done"
+        android:icon="@drawable/ic_done"
+        app:showAsAction="always"
+        android:title="@string/action_done"/>
+</menu>
